feat: merge gmBTC and Thena VIPs
